swapped out httpoison for req and everything is working
