+++ test/support/mocks.ex
@@ -9,5 +9,6 @@ Mox.defmock(Dotcom.Redis.Mock, for: Dotcom.Redis.Behaviour)
 Mox.defmock(Dotcom.Redix.Mock, for: Dotcom.Redix.Behaviour)
 Mox.defmock(Dotcom.Redix.PubSub.Mock, for: Dotcom.Redix.PubSub.Behaviour)
 
+Mox.defmock(CMS.Api.Mock, for: CMS.Api.Behaviour)
 Mox.defmock(MBTA.Api.Mock, for: MBTA.Api.Behaviour)
 Mox.defmock(OpenTripPlannerClient.Mock, for: OpenTripPlannerClient.Behaviour)
